+++ crates/common/types/block.rs
@@ -19,6 +19,7 @@ use ethrex_rlp::{
 };
 use ethrex_trie::Trie;
 use keccak_hash::keccak;
+use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use serde::{Deserialize, Serialize};
 
 use std::cmp::{Ordering, max};
@@ -233,11 +234,8 @@ impl BlockBody {
         // Recovering addresses is computationally expensive.
         // Computing them in parallel greatly reduces execution time.
         self.transactions
-            .iter()
-            .map(|tx| {
-                let t = (tx, tx.sender()?);
-                Ok(t)
-            })
+            .par_iter()
+            .map(|tx| Ok((tx, tx.sender()?)))
             .collect::<Result<Vec<(&Transaction, Address)>, secp256k1::Error>>()
     }
 }